Revert "refactor getIslandWidth into new helpers.ts (delete helpers.js)"
This reverts commit 2cd9c61af30be14263ff60065828b7a71a441e67.
